refactor(ext/node): consolidate node:fs modules (part 3) (#32525)
## Summary

- Inline `_fs_fchown.ts`, `_fs_ftruncate.ts`, `_fs_futimes.ts`,
`_fs_lchmod.ts`, `_fs_lchown.ts` into `fs.ts`
- Move promise variants into `internal/fs/promises.ts` and
`internal/fs/handle.ts`
- Consolidate tests from deleted test files into `fs_test.ts`
- Continues the series from #32502 and #32505

---------
