+++ ext/node/lib.rs
@@ -383,13 +383,8 @@ deno_core::extension!(deno_node,
     "_fs/cp/cp_sync.ts",
     "_fs/_fs_dir.ts",
     "_fs/_fs_exists.ts",
-    "_fs/_fs_fchown.ts",
     "_fs/_fs_fstat.ts",
-    "_fs/_fs_ftruncate.ts",
-    "_fs/_fs_futimes.ts",
     "_fs/_fs_glob.ts",
-    "_fs/_fs_lchmod.ts",
-    "_fs/_fs_lchown.ts",
     "_fs/_fs_lstat.ts",
     "_fs/_fs_lutimes.ts",
     "_fs/_fs_mkdir.ts",
