+++ ext/node/polyfills/internal/fs/promises.ts
@@ -15,8 +15,6 @@ import { promisify } from "ext:deno_node/internal/util.mjs";
 import * as constants from "ext:deno_node/_fs/_fs_constants.ts";
 import { copyFilePromise } from "ext:deno_node/_fs/_fs_copy.ts";
 import { cpPromise } from "ext:deno_node/_fs/_fs_cp.ts";
-import { lchmodPromise } from "ext:deno_node/_fs/_fs_lchmod.ts";
-import { lchownPromise } from "ext:deno_node/_fs/_fs_lchown.ts";
 import { lstatPromise } from "ext:deno_node/_fs/_fs_lstat.ts";
 import { lutimesPromise } from "ext:deno_node/_fs/_fs_lutimes.ts";
 import { mkdirPromise } from "ext:deno_node/_fs/_fs_mkdir.ts";
@@ -55,8 +53,11 @@ import {
 } from "ext:deno_node/internal/validators.mjs";
 import type { Buffer } from "node:buffer";
 import { primordials } from "ext:core/mod.js";
+import { op_node_lchmod, op_node_lchown } from "ext:core/ops";
+import { isMacOS } from "ext:deno_node/_util/os.ts";
+import { ERR_METHOD_NOT_IMPLEMENTED } from "ext:deno_node/internal/errors.ts";
 
-const { Error, PromisePrototypeThen } = primordials;
+const { Error, PromisePrototypeThen, PromiseReject } = primordials;
 
 // -- access --
 
@@ -205,7 +206,34 @@ const chownPromise = promisify(chown) as (
   gid: number,
 ) => Promise<void>;
 
-// -- link --
+const lchmodPromise: (
+  path: string | Buffer | URL,
+  mode: number,
+) => Promise<void> = !isMacOS
+  ? () => PromiseReject(new ERR_METHOD_NOT_IMPLEMENTED("lchmod()"))
+  : (path: string | Buffer | URL, mode: number) => {
+    path = getValidatedPathToString(path);
+    mode = parseFileMode(mode, "mode");
+    return op_node_lchmod(path, mode);
+  };
+
+function lchown(
+  path: string | Buffer | URL,
+  uid: number,
+  gid: number,
+  callback: CallbackWithError,
+) {
+  callback = makeCallback(callback);
+  path = getValidatedPathToString(path);
+  validateInteger(uid, "uid", -1, kMaxUserId);
+  validateInteger(gid, "gid", -1, kMaxUserId);
+
+  PromisePrototypeThen(
+    op_node_lchown(path, uid, gid),
+    () => callback(null),
+    callback,
+  );
+}
 
 function link(
   existingPath: string | Buffer | URL,
@@ -222,12 +250,16 @@ function link(
   );
 }
 
+const lchownPromise = promisify(lchown) as (
+  path: string | Buffer | URL,
+  uid: number,
+  gid: number,
+) => Promise<void>;
+
 const linkPromise = promisify(link) as (
   existingPath: string | Buffer | URL,
   newPath: string | Buffer | URL,
 ) => Promise<void>;
-
-// -- unlink --
 
 function unlink(
   path: string | Buffer | URL,
